refactor(summaries): mode-branch SummariesPage; multi-label intact
