+++ src/pages/summaries/SummariesPageSingle.tsx
@@ -1,0 +1,14 @@
+export function SummariesPageSingle() {
+  return (
+    <div className="flex-1 overflow-auto bg-canvas">
+      <div className="max-w-[960px] mx-auto px-12 py-12 text-on-canvas">
+        <h1 className="font-serif font-medium text-[32px] text-paper tracking-[-0.018em] m-0 mb-2">
+          Summaries
+        </h1>
+        <p className="font-serif text-on-surface text-[14px]">
+          Single-label master-detail UI coming online — Tasks 15–26 will fill this in.
+        </p>
+      </div>
+    </div>
+  )
+}
